format + moved away s2gos/client/gui/create.py
