+++ s2gos/client/client.py
@@ -15,6 +15,7 @@ from s2gos.common.models import (
     ProcessRequest,
     Results,
 )
+
 from .transport import DefaultTransport, Transport
