+++ s2gos/client/transport.py
@@ -17,7 +17,6 @@ from s2gos.client.config import ClientConfig
 from s2gos.client.defaults import DEFAULT_SERVER_URL
 from s2gos.client.exceptions import ClientException
 
-
 logger = getLogger("s2gos")
