+++ s2gos/server/app.py
@@ -2,8 +2,8 @@
 #  Permissions are hereby granted under the terms of the Apache 2.0 License:
 #  https://opensource.org/license/apache-2-0.
 
-import time
 import logging
+import time
 from typing import Awaitable, Callable
 
 from fastapi import FastAPI, Request, Response
